Change binaries. Replaced jar manually, check commit description.
The last commit was crashing with hermes. This commit now uses all the compiled files from "react-native": "npm:react-native-tvos@0.68.2-3”. The only thing which is changed is the classes.jar inside the aar file. We use this dirty approach to fix ARD-768. We hope to fix this later in a better, if with upgrading react-native hermes will be compiled by ourself (will happen with >0.69.